+++ config/config.exs
@@ -10,7 +10,8 @@ if config_env() == :dev do
   config :esbuild,
     version: "0.14.41",
     default: [
-      args: ~w(js/app.js --bundle --target=es2020 --outdir=../dist/js),
+      args:
+        ~w(js/app.js --bundle --minify --sourcemap=external --target=es2020 --outdir=../dist/js),
       cd: Path.expand("../assets", __DIR__),
       env: %{"NODE_PATH" => Path.expand("../deps", __DIR__)}
     ]
@@ -18,7 +19,7 @@ if config_env() == :dev do
   config :dart_sass,
     version: "1.54.5",
     default: [
-      args: ~w(--load-path=node_modules css/app.scss ../dist/css/app.css),
+      args: ~w(--load-path=node_modules --no-source-map css/app.scss ../dist/css/app.css),
       cd: Path.expand("../assets", __DIR__)
     ]
 end